scraper/requirements.txt
playwright==1.41.0
beautifulsoup4>=4.12.0
requests>=2.31.0
supabase>=2.3.0
google-genai>=1.0.0
python-dateutil>=2.8.2
lxml>=5.0.0
python-dotenv>=1.0.0
